♻️ refactor :   컴포넌트 구조 컨벤션처럼
