+++ src/app/layout.tsx
@@ -1,9 +1,7 @@
 import './globals.css';
 
-import { Suspense } from 'react';
-
-import { JsonLD } from '@/components';
-import { Analytics } from '@/components/Analytics';
+import Analytics from './Analytics';
+import JsonLD from './JsonLD';
 import { OverlayProvider } from '@/components/Overlay/OverlayProvider';
 import METADATA from '@/constants/meta';
 import QueryProvider from '@/provider/QueryProvider';
@@ -17,8 +15,8 @@ export const metadata: Metadata = METADATA;
 export default function RootLayout({ children }: PropsWithChildren) {
   return (
     <html lang="ko">
-      <body className="flex justify-center w-screen touch-none bg-slate-100">
-        <div className="w-full overflow-scroll bg-white max-w-layout text-primary">
+      <body className="flex w-screen touch-none justify-center bg-slate-100">
+        <div className="w-full max-w-layout overflow-scroll bg-white text-primary">
           <Analytics />
           <JsonLD />
           <QueryProvider>
