--- src/components/Analytics/index.ts
@@ -1,1 +1,0 @@
-export { default as Analytics } from './Analytics';
