--- src/components/JsonLD/index.ts
@@ -1,1 +1,0 @@
-export { default as JsonLD } from './JsonLD';
